- reworking some shitty backend stuff and registerung listeners and commands

+++ src/main/java/de/chaos/mc/banksystem/utils/ITransaktionInterface.java
@@ -5,7 +5,7 @@ import java.util.UUID;
 
 public interface ITransaktionInterface {
 
-    public TransaktionLogsDAO newTransaktion(UUID uuid, UUID target, long amount);
+    public TransaktionLogsDAO newTransaktion(UUID uuid, UUID target, int amount);
 
     public List<TransaktionLogsDAO> getLastTransaktions(UUID uuid);
 }
